Added MOVTransaction and HistoryFragment
MOVTransaction is just an object type mirroring the one on Firebase. it has
    String from;
    String from_name;
    String to;
    String to_name;
    int amount;
    String currency;
    long time;
as fields. It has a toString function that returns a sentence for HistoryFragment to use.

fragment_history.xml simply has a textfield saying "Recent Transactions" and a ListView that can be filled.

History Fragment just uses firebase to fill the ListView from the transactions recorded there.